+++ sote/libsote/world-generator.lua
@@ -107,7 +107,7 @@ local function glaciers()
 
 	table.insert(prof_output, { profile_and_get(function() require "libsote.soils.gen-bias-matrix".run(wg.world) end, "gen-bias-matrix", 1) })
 	table.insert(prof_output, { profile_and_get(function() require "libsote.soils.gen-parent-material".run(wg.world) end, "gen-parent-material", 1) })
-	table.insert(prof_output, { profile_and_get(function() require "libsote.glacial-formation".run(wg.world) end, "glacial-formation", 1) })
+	table.insert(prof_output, { profile_and_get(function() require "libsote.glaciation.glacial-formation".run(wg.world) end, "glacial-formation", 1) })
 
 	log_profiling_data(prof_output, "glaciers")
 end
